@@ -18,11 +18,11 @@ sys.path.insert(0, os.path.abspath('../..'))
 # -- Project information -----------------------------------------------------
 
 project = 'PyFHD'
-copyright = '2022, Joel Dunstan and Nichole Barry'
-author = 'Joel Dunstan and Nichole Barry'
+copyright = '2022, Joel Dunstan, Jack Line and Nichole Barry'
+author = 'Joel Dunstan, Jack Line and Nichole Barry'
 
 # The full version, including alpha/beta/rc tags
-release = '0.1'
+release = '1.0'
 
 
 # -- General configuration ---------------------------------------------------
@@ -37,6 +37,7 @@ extensions = [
               'sphinx.ext.napoleon',
               'sphinx.ext.todo',
               'sphinxarg.ext',
+              'myst_parser'
              ]
 
 # Display todos by setting to True
@@ -62,3 +63,9 @@ html_theme = 'sphinx_rtd_theme'
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
 html_static_path = ['_static']
+
+source_suffix = {
+    '.rst': 'restructuredtext',
+    '.txt': 'markdown',
+    '.md': 'markdown',
+}
