@@ -30,7 +30,7 @@ def test_weight_invert_three(data_dir):
                                                               'visibility_grid_input_weights_3.npy', 
                                                               'visibility_grid_output_result_3.npy',
                                                               'visibility_grid_input_abs_3.npy')
-    result = weight_invert(weights, threshold = threshold)
+    result = weight_invert(weights, threshold = threshold, use_abs = abs)
     assert np.array_equal(result, expected_result)
 
 def test_weight_invert_four(data_dir):
